[AWIBOF-3825] apply code review comments
Former-commit-id: f81ee9e505a7cd0042d44c4491dd9d78c8cfdf55

+++ src/array/rebuild/rebuild_logger.h
@@ -48,7 +48,7 @@ public:
     void SetPartitionRebuildStart(string partName);
     void SetResult(string result);
     void WriteLog(void);
-    uint32_t rebuiltSegCnt = 0;
+    uint64_t rebuiltSegCnt = 0;
 
 private:
     string array;
